quantum_cryptography/protocols/test_quantum_superposition_entanglement_example.py
import unittest
from qiskit import Aer, transpile, assemble, execute
from quantum_superposition_entanglement_example import quantum_superposition_entanglement

class TestQuantumSuperpositionEntanglementExample(unittest.TestCase):

    def test_quantum_superposition_entanglement(self):
        qc = quantum_superposition_entanglement()
        
        # Simulate the circuit
        backend = Aer.get_backend('qasm_simulator')
        tqc = transpile(qc, backend)
        qobj = assemble(tqc)
        result = execute(qc, backend).result()
        counts = result.get_counts()
        
        # Check if the result contains the expected entangled outcome
        # Expecting equal probabilities for '00' and '11' states due to entanglement
        self.assertIn('00', counts)
        self.assertIn('11', counts)
        self.assertAlmostEqual(counts['00'], counts['11'], delta=100)

if __name__ == '__main__':
    unittest.main()
